+++ src/Paket.Core/RemoteDownload.fs
@@ -149,9 +149,6 @@ let downloadRemoteFiles(remoteFile:ResolvedSourceFile,destination) = async {
             ) |> Async.Parallel
         task |> Async.RunSynchronously |> ignore
 
-        // GIST currently does not support zip-packages, so now this fetches all files separately.
-        // let downloadUrl = sprintf "https://gist.github.com/%s/%s/download" remoteFile.Owner remoteFile.Project //is a tar.gz
-
     | SingleSourceFileOrigin.GitHubLink, Constants.FullProjectSourceFileName -> 
         let fi = FileInfo(destination)
         let projectPath = fi.Directory.FullName
@@ -175,12 +172,13 @@ let downloadRemoteFiles(remoteFile:ResolvedSourceFile,destination) = async {
     | SingleSourceFileOrigin.HttpLink(origin), _ ->
         let url = origin + remoteFile.Commit
         let authentication = auth remoteFile.AuthKey url
+        let targetFolder = FileInfo(destination).Directory
+        if not targetFolder.Exists then
+            targetFolder.Create()
         match Path.GetExtension(destination).ToLowerInvariant() with
         | ".zip" ->
-            let targetFolder = FileInfo(destination).Directory.FullName
-            CleanDir targetFolder
             do! downloadFromUrl(authentication, url) destination
-            ZipFile.ExtractToDirectory(destination, targetFolder)
+            ZipFile.ExtractToDirectory(destination, targetFolder.FullName)
         | _ -> do! downloadFromUrl(authentication, url) destination
 }
 
@@ -201,31 +199,24 @@ let DownloadSourceFiles(rootPath, groupName, force, sourceFiles:ModuleResolver.R
             CleanDir destinationDir
 
         (versionFile, version), sources)
-    |> List.map (fun ((versionFile, version), sources) ->
-        async {
-            let! downloaded =
-                sources
-                |> List.map (fun (_, (destination, source)) ->
-                    async {
-                        let exists =
-                            if destination.EndsWith Constants.FullProjectSourceFileName then
-                                let di = FileInfo(destination).Directory
-                                di.Exists && FileInfo(Path.Combine(di.FullName, Constants.PaketVersionFileName)).Exists
-                            else
-                                File.Exists destination
+    |> List.iter (fun ((versionFile, version), sources) ->
+        sources
+        |> List.iter (fun (_, (destination, source)) ->
+            let exists =
+                if destination.EndsWith Constants.FullProjectSourceFileName then
+                    let di = FileInfo(destination).Directory
+                    di.Exists && FileInfo(Path.Combine(di.FullName, Constants.PaketVersionFileName)).Exists
+                else
+                    File.Exists destination
 
-                        if not force && exists then
-                            verbosefn "Sourcefile %O is already there." source
-                        else 
-                            tracefn "Downloading %O to %s" source destination
-                            do! downloadRemoteFiles(source,destination)
-                    })
-                |> Async.Parallel
+            if not force && exists then
+                verbosefn "Sourcefile %O is already there." source
+            else 
+                tracefn "Downloading %O to %s" source destination
+                Async.RunSynchronously <| downloadRemoteFiles(source,destination))
 
-            if File.Exists(versionFile.FullName) then
-                if not <| File.ReadAllText(versionFile.FullName).Contains(version) then
-                    File.AppendAllLines(versionFile.FullName, [version])
-            else
+        if File.Exists(versionFile.FullName) then
+            if not <| File.ReadAllText(versionFile.FullName).Contains(version) then
                 File.AppendAllLines(versionFile.FullName, [version])
-        })
-    |> Async.Parallel
+        else
+            File.AppendAllLines(versionFile.FullName, [version]))